edit booking

+++ app/angular/AngularView/src/app/booking-confirmation/booking-confirmation.component.ts
@@ -1,8 +1,14 @@
-import { MatSnackBar } from '@angular/material/snack-bar';
-import { Booking } from './../model-service/bookings/bookings';
-import { ActivatedRoute } from '@angular/router';
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
+
+import { Booking } from './../model-service/bookings/bookings';
 import { BookingsService } from '../model-service/bookings/bookings.service';
+import { BookedItem } from '../model-service/items/items';
+
+import { ConfirmationDialogComponent } from '../confirmation-dialog/confirmation-dialog.component';
+
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-booking-confirmation',
@@ -12,39 +18,70 @@ import { BookingsService } from '../model-service/bookings/bookings.service';
 export class BookingConfirmationComponent implements OnInit {
   booking: Booking;
   token: string;
+  tableColumnsBookedItems: string[] = ['name', 'quantity', 'status'];
+  bookedItems: BookedItem[];
+  responded: boolean;
 
   constructor(
     private route: ActivatedRoute,
     private bookingsService: BookingsService,
-    private snackbar: MatSnackBar
+    private snackbar: MatSnackBar,
+    public dialog: MatDialog
   ) { }
 
   ngOnInit(): void {
     this.route.queryParams.subscribe(params => {
-      this.token = params['token'];
-  });
+      this.token = params.token;
+    });
     console.log(this.token);
     this.bookingsService.getBookingByToken(this.token).subscribe(
       (booking: Booking) => {
         this.booking = booking;
+        this.bookingsService.getBookedItemsByBooker(booking.id).subscribe(
+          (data: BookedItem[]) => {
+            this.bookedItems = data;
+          });
       }
     );
+  }
+
+  getStatus(code: string) {
+    switch (code) {
+      case 'PEN':
+        return 'Pending';
+      case 'EVA':
+        return 'Evaluating';
+      case 'PRO':
+        return 'Processed';
+      case 'UNC':
+        return 'Unconfirmed';
+      default:
+        return 'Unknown';
+    }
   }
 
   confirm_booking() {
     this.bookingsService.confirmBookingByToken(this.token).subscribe(
       (success: any) => {
-        this.snackbar.open('Booking confirmed', 'OK', {duration: 5000, });
+        this.responded = true; // order matters here, otherwise user can't do any actions then gg
+        location.reload(); // force reload
+        this.snackbar.open('Booking confirmed', 'OK', { duration: 5000, });
       }
     );
   }
 
   delete_booking() {
-    this.bookingsService.deleteBookingByToken(this.token).subscribe(
-      (success: any) => {
-        this.snackbar.open('Booking deleted', 'OK', {duration: 5000, });
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, { data: 'this booking' });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result.event === 'yes') {
+        this.bookingsService.deleteBookingByToken(this.token).subscribe(
+          (success: any) => {
+            this.responded = true;
+            location.reload(); // force reload
+            this.snackbar.open('Booking deleted', 'OK', { duration: 5000, });
+          }
+        );
       }
-    );
+    });
   }
-
 }
